+++ kotlinx/coroutines/core/src/commonMain/kotlin/kotlinx/coroutines/universal/Dispatchers.kt
@@ -1,9 +1,0 @@
-package kotlinx.coroutines.universal
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.MainCoroutineDispatcher
-
-expect object Dispatchers {
-    val Universal: CoroutineDispatcher
-    val Main: MainCoroutineDispatcher
-}