+++ kotlinx/coroutines/core/src/jvmMain/kotlin/kotlinx/coroutines/universal/Dispatchers.kt
@@ -1,9 +1,0 @@
-package kotlinx.coroutines.universal
-
-import kotlinx.coroutines.*
-import kotlinx.coroutines.Dispatchers
-
-actual object Dispatchers {
-    actual val Universal: CoroutineDispatcher = Dispatchers.Default
-    actual val Main: MainCoroutineDispatcher = Dispatchers.Main
-}