+++ live/react/src/main/kotlin/hooks.kt
@@ -4,11 +4,10 @@ import live.Live
 import react.*
 
 fun <S> useLive(live: Live<S>): S {
-    val (state, setState) = useState(live.value)
-    rawUseEffect({
-        val listener = live.watch { setState(it) }
-        val cleanup: () -> Unit = { listener.stop() }
-        cleanup
-    }, emptyArray())
+    var state by useState(live.value)
+    useEffectOnce {
+        val listener = live.watch { state = it }
+        cleanup { listener.stop() }
+    }
     return state
 }