+++ viewmodel/core/src/jvmTest/kotlin/logging/ViewModelLogging.kt
@@ -1,17 +1,12 @@
 @file:Suppress("PackageDirectoryMismatch")
 
-import logging.TodoViewModel.Intent
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.asCoroutineDispatcher
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
-import logging.*
-import test.asyncTest
-import java.util.concurrent.Executors
-import kotlin.test.AfterTest
-import kotlin.test.BeforeTest
+import kotlinx.coroutines.runTest
+import logging.ConsoleAppender
+import logging.Logging
+import logging.TodoViewModel
+import logging.TodoViewModel.Intent
+import viewmodel.expect
 import kotlin.test.Test
 
 class ViewModelLogging {
@@ -21,9 +16,9 @@ class ViewModelLogging {
     }
 
     @Test
-    fun should_print_logging_output() = asyncTest {
+    fun should_print_logging_output() = runTest {
         val vm = TodoViewModel()
         delay(50)
-        vm.post(Intent.ReInit)
+        vm.expect(Intent.ReInit).toGoThrough(TodoViewModel.State.Init)
     }
 }