+++ viewmodel/react/src/main/kotlin/viewmodel/StateExtensions.kt
@@ -1,0 +1,5 @@
+package viewmodel
+
+import useLive
+
+fun <S> ViewModel<*,S>.asState() = useLive(ui)